requirements/dev.txt
autopep8>=1.5.4
migra>=3.0.16,<4.0.0
mypy>=0.800,<0.810
pika-stubs==0.1.3
pycodestyle>=2.6.0
pydocstyle>=5.1.1
pylint>=2.6.0
sqlbag>=0.1.15,<0.2.0
